StudentList & Student Commit
finished the loadStudentList and saveStudentList in the StudentList
class.
Added setters and getters for the Student class...thought those were in
there before...guess not?

Made changes to use constants from Amits constants file.
Did change the path in there to point to the .\\data\Students.txt  and
.\\data\\CourseList.txt and it works with the menus and reading the
files on mine fine.