Deleted source/data/ & update Constants
Amit and I talked tonight and discovered that \ (backslashes) are not
able to read on the MAC the same as Windows, so changed paths to use /
(forward slashes) and they seem to work.  I was able to show Amit that
the menu system and file reading/writing worked using the
./data/Student.txt directory structure and using the Constants that he
has defined.