Registrar: saveRegistration,saveCoursesAll, and myCourseSchedule
I believe we are ready to add this to the menu to register for a course
by calling registerForCourse(int studentID, String courseID)
and then to show the students own course schedule
myCourseSchedule(int studentID).

There are several support methods for loading, saving registrations and
incrementing the counters for enrollment in the courses and also for
saving the updated courses ArrayList back out to the CourseList.txt
file.

I still have to work on dropping a course, but figured we had better get
this into the menu system and get testing on it.